refactor(kda): move forward logic from base class to KDAAttnBackend
Align with GPU sglang architecture where MambaAttnBackendBase only
provides metadata infra and state tracking, while KDAAttnBackend owns
the full forward flow (QKV split, gate/beta reshape, mode dispatch,
kernel calls, state writeback).

- LinearRecurrentAttnBackend: now metadata + pytree infra only
- KDAAttnBackend: __call__, _forward_extend, _forward_decode,
  _split_qkv, _reshape_gate, _reshape_beta
- _split_qkv simplified to 2 formats (tuple, flat 2D) vs prior 4
